requirements.txt
opencv-python
numpy
yaml
insightface
ultralytics
mediapipe
pillow
tqdm
tk
sqlite3 